Update aibot_twitter.py
This function sends a direct message using the Twitter API v2. It constructs the appropriate data for the message and makes a POST request to the Twitter API using the requests library. The BEARER_TOKEN is used to authenticate the request. If the message is sent successfully, it prints a success message. If there is an error, it prints an error message with the status code.